Faster script domain resolution
Probably the most called WM functions are domain-resolution related:
every time a DOMContentLoaded event fires in the browser, the document URI
is tested against every enabled script's set of include and exclude rules.
Even a slight improvement in the implementation might result in an
appreciable gain. That's what we do here, in storing the regular expression
associated to an include/exclude mask.
There might be memory gains as well as the old mechanism was creating a lot
of single-use RegExp objects.